django-app/requirements.txt
Django==4.2.0
gunicorn==20.1.0
django-environ==0.10.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==3.14.0
pillow==9.5.0
whitenoise==6.4.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-widget-tweaks==1.4.12
django-cleanup==7.0.0
psycopg2-binary==2.9.6
uvicorn==0.22.0
python-json-logger==2.0.7
sentry-sdk==1.25.1
django-health-check==3.17.0
